Using framework web2py to construct marreta
** This is not a functional version, but the commit is necessary to us **

This is the best way to create marreta system.
Please, take a look on this. To check everything that I did here,
please follow those steps:
A. Database:
1. Install the MySQL
2. Create a table with name marreta
3. Edit {web2py}/Contents/Resources/applications/marreta/private/appconfig.ini

B. Marreta
1. Go to http://web2py.com and Download web2py for your OS
2. Unzip/Uncompress the web2py
3. Go to directory Contents/Resources/applications/
4. Do the git clone for this review
5. Run the web2py.py (Contents/Resources/)
6. Got to http://127.0.0.1/marreta/
7. Select Marreta Administration
8. Default login: admin@marreta.org / abc1234

On this review I create two things to demonstrate how is
easy for us use this framework:

Check please the controller/admin.py - This is all our administration
interface (WIP), for sure this is only a seed and need to be improved. But it toke
from me just 1 hour to contruct it.
Included on this time I did the database design that you can check out in
modes/db.py (including crypto for passwords)
The views, interfaces are available on views/admin/*.html (also easy to work)

Please, do your comments and let keep in touch.
IMHO, with a good framework we can skip problems about security stufs,
even a lot of stuffs non-technicals ;)

Change-Id: I9c1b4e7bac29ead3a651e50fc15f98815655beba