Database re-design and refactor
Most of the DB "Constraints" will be done by Application.
Be aware that the only cascade that was activated is on
ConfigurationItem remove.
If a CI is removed, all username/password records of that CI
on ServerCredentials will be dropped.
Refactor was done in controllers/adminpy because of the database changes.

TODO: Probably in future, the field "Password" may be adapted to support
encrypted password.
Still working in which technology will be used.
But wont make a differce as it will be seted by marreta engine.
TODO: Form should not show options to fillup Server Status,
LastServerChk or AddedDate.
This should be filled by appliacion itself.

Change-Id: I74cf2b9842a7f02c947ef9e22c0eb90df748ee46